Move visit_dyn_array to its own file, so it can be shared with DynArrayB.
